+++ src/main/java/com/transactions/TransactionSaver.java
@@ -1,24 +1,17 @@
 package com.transactions;
 
-import com.jsonUsage.ReadJSONFileFromPath;
+import com.jsonUsage.JsonConfig;
 import com.jsonUsage.WrongJsonPathException;
 
 import java.io.*;
 import java.util.*;
 import java.lang.*;
 
-import static com.jsonUsage.ReadInputArgs.PATH_TO_JSON;
-
 public class TransactionSaver {
-    static ReadJSONFileFromPath readJSONFileFromPath = new ReadJSONFileFromPath();
     static String finalTransactionFilePath;
 
     static {
-        try {
-            finalTransactionFilePath = readJSONFileFromPath.getJsonFilePath(PATH_TO_JSON);
-        } catch (WrongJsonPathException e) {
-            e.printStackTrace();
-        }
+        finalTransactionFilePath = JsonConfig.getTransactionsPath();
     }
 
     public static String DEFAULT_FILE = finalTransactionFilePath;
