+++ transactions.csv
@@ -1,0 +1,2 @@
+ID;Name;Description;Date;Currency;Price
+1;Test trans;foo;Thu Jun 10 22:54:55 CEST 2021;$;10.0
